# swagger.yaml
openapi: 3.0.0
info:
  title: Training
  contact: {}
  version: '1.0'
servers:
- url: http://localhost:3000
  variables: {}
paths:
...
  /post/editPost/b24041d2-c098-4f11-bb2b-5f3b93b5cc58:
    patch:
      tags:
      - Post
      summary: Edit post controller
      operationId: Editpostcontroller
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/EditpostcontrollerRequest'
              - example:
                  title: post 11
                  desc: post 11
                  verification: '0'
            example:
              title: post 11
              desc: post 11
              verification: '0'
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
  /post/deleteOne:
    delete:
      tags:
      - Post
      summary: Delete One Post
      operationId: DeleteOnePost
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
  /post/viewOne/d0116b92-0129-4e9f-8159-7b9f14dc573f:
    get:
      tags:
      - Post
      summary: View one post
      operationId: Viewonepost
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
  schemas:
...
    EditpostcontrollerRequest:
      title: EditpostcontrollerRequest
      required:
      - title
      - desc
      - verification
      type: object
      properties:
        title:
          type: string
        desc:
          type: string
        verification:
          type: string
      example:
        title: post 11
        desc: post 11
        verification: '0'
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...